+++ ui/app/src/components/NewExperimentNext/data/types.ts
@@ -1254,6 +1254,12 @@ export const dataPhysic: Record<KindPhysic, Definition> = {
         value: 9,
         helperText: 'The process signal value',
       },
+      recoverCmd: {
+        field: 'text',
+        label: 'Recover Command',
+        value: '',
+        helperText: 'The command to be run when recovering experiment',
+      },
     },
   },
   StressChaos: {